GOS connected to basketgame
The logic will work to give 1,2,3 stars according to the number of bananas added to the basket. If the number exactly matches the requiredApples (7) it will give 3 stars, if the number varies +-1, it will give only 2 starts, if the number varies +-2 it will only give 1 star. If the number varies >=3 it will give 0 stars. The reaction time is also being calculated and passed into the SubmitGameResults() method with the gameDuration.